requirements.txt
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-gis==0.14
